refactor #258: Split raft_test.rs into modular test structure with category organization
- Moved monolithic raft_test.rs → raft_test/mod.rs (module root)
- Created 4 specialized test modules organized by test category:
  * leader_change_tests.rs (Category E1: Notification Tests, 4 tests)
  * leader_discovered_tests.rs (Category B6+E: RoleEvent Handling, 5 tests)
  * raft_comprehensive_tests.rs (Categories A,C,D,F,G,H: Core Raft, 78 tests)
  * drain_based_batch_architecture_tests.rs (Category I: Batch Architecture, 14 tests)

- Added file-level headers with metadata (original location, test count, category)
- Fixed all imports (super::super:: → crate::)
- Updated mod.rs with full Category I documentation (I1-I4 subcategories)
- All 115 tests passing, compilation clean
- Git history preserved via git mv raft_test.rs raft_test/mod.rs
