+++ d-engine-core/src/raft_test/raft_comprehensive_tests.rs
@@ -1,519 +1,27 @@
-//! # Comprehensive Raft Unit Test Scenarios
+//! # Comprehensive Raft Tests
+//!
+//! **Original location**: raft_test.rs L517-2528
+//! **Test count**: 78 tests
+//! **Categories**: A (Role Transitions), B (RoleEvent Handling), C (Leader Init),
+//!                  D (Leadership Verification), F (Event Queue), G (Error Cases),
+//!                  H (Bootstrap & Joining)
+//!
+//! This is the largest test module covering core Raft consensus behavior:
+//! - A: Role state machine transitions (25 tests)
+//! - B: RoleEvent processing (6 tests within this module)
+//! - C: Leader initialization of peer state (12 tests)
+//! - D: Leadership verification via noop entry (9 tests)
+//! - F: Event loop priority and election timeout (12 tests)
+//! - G: Error cases and edge cases (3 tests)
+//! - H: Node joining and bootstrap (17 tests)
+//!
+//! **Key Testing Approach**:
+//! - Uses MockBuilder for Raft instance creation
+//! - tokio::time::pause() for deterministic timing
+//! - Mock membership and log storage
+//! - Listener notification verification
 //!
-//! This module contains comprehensive unit tests for the Raft consensus implementation.
-//! These tests follow the Raft paper (https://raft.github.io) and best practices for
-//! consensus algorithm testing.
-//!
-//! ## Test Organization
-//!
-//! Tests are organized into the following categories:
-//!
-//! ### A. Role Transition Tests (State Machine Validation)
-//!
-//! These tests verify the correctness of the Raft role state machine - that transitions
-//! between states (Follower, Candidate, Leader, Learner) follow Raft protocol rules.
-//!
-//! **Key Invariants:**
-//! - Only specific role transitions are valid
-//! - Invalid transitions are rejected with errors
-//! - State machine rules prevent protocol violations
-//!
-//! #### A1. Valid Transition Paths
-//! - **A1.1**: Follower → Candidate → Leader → Follower (normal election and step-down)
-//! - **A1.2**: Follower → Candidate → Follower (election failure)
-//! - **A1.3**: Follower → Candidate → Leader (normal election success)
-//! - **A1.4**: Candidate → Follower (discovered higher term)
-//! - **A1.5**: Candidate → Candidate (invalid: cannot re-become Candidate)
-//! - **A1.6**: Leader → Follower (valid: step down on higher term)
-//! - **A1.7**: Leader → Candidate (invalid: cannot become Candidate)
-//! - **A1.8**: Leader → Leader (invalid: already Leader)
-//! - **A1.9**: Learner → Follower (valid: learner promotion path)
-//! - **A1.10**: Learner → Candidate (invalid: learners cannot vote)
-//! - **A1.11**: Learner → Leader (invalid: learners are non-voting)
-//! - **A1.12**: Learner → Learner (invalid: cannot re-become Learner)
-//!
-//! #### A2. Follower Role Entry/Exit
-//! - **A2.1**: BecomeFollower with leader_id=Some(X) - discovers known leader
-//! - **A2.2**: BecomeFollower with leader_id=None - no leader known
-//! - **A2.3**: Follower state reset: voted_for should be cleared
-//! - **A2.4**: Follower state reset: term should not change on transition
-//!
-//! #### A3. Candidate Role Entry/Exit
-//! - **A3.1**: BecomeCandidate - increments term
-//! - **A3.2**: BecomeCandidate - sets voted_for to self
-//! - **A3.3**: BecomeCandidate - clears leader info (no leader in candidate state)
-//! - **A3.4**: Candidate timeout triggers election (tested with tokio fake timer)
-//!
-//! #### A4. Leader Role Entry/Exit
-//! - **A4.1**: BecomeLeader - marks vote as committed
-//! - **A4.2**: BecomeLeader - initializes peer tracking (next_index, match_index)
-//! - **A4.3**: BecomeLeader - initializes cluster metadata cache
-//! - **A4.4**: BecomeLeader - sends noop entry for leadership verification
-//! - **A4.5**: Leadership verification succeeds - leader established
-//! - **A4.6**: Leadership verification fails - auto-downgrades to Follower
-//! - **A4.7**: Single-node cluster - skips peer initialization
-//!
-//! #### A5. Learner Role Entry/Exit
-//! - **A5.1**: BecomeLearner - non-voting state (no leader info)
-//! - **A5.2**: Learner promotion to Follower
-//! - **A5.3**: Multiple learners in cluster (tracked separately from voters)
-//!
-//! ### B. RoleEvent Handling Tests (Individual Event Processing)
-//!
-//! These tests verify that each RoleEvent variant is processed correctly and sends
-//! appropriate notifications to listeners.
-//!
-//! #### B1. BecomeFollower Event
-//! - **B1.1**: Valid transition and state reset
-//! - **B1.2**: leader_change notification sent with correct leader_id and term
-//! - **B1.3**: voted_for reset when becoming Follower
-//!
-//! #### B2. BecomeCandidate Event
-//! - **B2.1**: Valid transition from Follower to Candidate
-//! - **B2.2**: leader_change notification sent with None (no leader)
-//!
-//! #### B3. BecomeLeader Event
-//! - **B3.1**: Valid transition from Candidate to Leader
-//! - **B3.2**: leader_change notification sent with self node_id
-//! - **B3.3**: Peer next_index initialized to last_entry_id + 1
-//! - **B3.4**: Peer match_index initialized to 0
-//! - **B3.5**: Cluster metadata cache populated
-//! - **B3.6**: Leadership verification via noop entry
-//! - **B3.7**: Auto-downgrade to Follower on verification failure
-//!
-//! #### B4. BecomeLearner Event
-//! - **B4.1**: Valid transition from Follower to Learner
-//! - **B4.2**: leader_change notification sent with None
-//!
-//! #### B5. NotifyNewCommitIndex Event
-//! - **B5.1**: Notification broadcast to all registered listeners
-//! - **B5.2**: Correct new_commit_index passed to listeners
-//! - **B5.3**: Multiple commit listeners receive notification
-//!
-//! #### B6. LeaderDiscovered Event
-//! - **B6.1**: Notification sent to leader_change listeners
-//! - **B6.2**: Does NOT change current role (no state transition)
-//! - **B6.3**: Multiple listeners receive notification
-//!
-//! #### B7. ReprocessEvent Event
-//! - **B7.1**: Event re-queued to event_rx for reprocessing
-//!
-//! ### C. Leader Initialization Tests (Peer State Setup)
-//!
-//! These tests verify that newly elected leaders correctly initialize peer state
-//! for log replication. This is critical for correctness as incorrect initialization
-//! can lead to divergent logs.
-//!
-//! #### C1. Peer Index Initialization (Raft §5.3)
-//! - **C1.1**: Single peer - next_index = last_entry_id + 1, match_index = 0
-//! - **C1.2**: Two peer cluster - both peers initialized correctly
-//! - **C1.3**: Three peer cluster - all peers initialized correctly
-//! - **C1.4**: Five peer cluster - all peers initialized correctly
-//! - **C1.5**: Match index starts at 0 for all peers
-//! - **C1.6**: Learner peers also initialized (tracked separately)
-//! - **C1.7**: Mixed voter/learner cluster initialization
-//!
-//! #### C2. Cluster Metadata Caching (Hot Path Optimization)
-//! - **C2.1**: replication_peers() called to populate cache
-//! - **C2.2**: voters() called for quorum calculation
-//! - **C2.3**: Cache includes all active peers
-//! - **C2.4**: Learners tracked separately from voters
-//! - **C2.5**: Single-node cluster skips peer initialization
-//!
-//! ### D. Leadership Verification Tests (Raft §5.4.2)
-//!
-//! According to Raft paper, leader must append entry with its own term before
-//! committing entries from previous terms. These tests verify this critical behavior.
-//!
-//! #### D1. Leadership Verification Success Path
-//! - **D1.1**: Noop entry appended with current term
-//! - **D1.2**: Majority of peers respond positively
-//! - **D1.3**: Leader maintains role when verification succeeds
-//!
-//! #### D2. Leadership Verification Failure Path
-//! - **D2.1**: Majority of peers timeout/fail to respond
-//! - **D2.2**: Leader auto-downgrades to Follower
-//! - **D2.3**: Single-node cluster always succeeds
-//!
-//! #### D3. Quorum Calculation Edge Cases
-//! - **D3.1**: Exactly majority quorum (3-node cluster, 2 responses)
-//! - **D3.2**: Minority failure (5-node cluster, 3+ responses needed)
-//! - **D3.3**: Network partition - minority side loses leadership
-//!
-//! ### E. Notification and Listener Tests
-//!
-//! These tests verify the notification system works correctly for business logic
-//! to react to Raft state changes.
-//!
-//! #### E1. Leader Change Notifications
-//! - **E1.1**: Leader discovered - notification sent with leader_id
-//! - **E1.2**: No leader (Candidate/Learner state) - notification with None
-//! - **E1.3**: Multiple listeners receive all notifications
-//! - **E1.4**: Watch channel deduplication (identical notifications)
-//!
-//! #### E2. New Commit Notifications
-//! - **E2.1**: Notification sent with correct commit index
-//! - **E2.2**: Multiple listeners receive notification
-//!
-//! #### E3. Role Transition Notifications (Test-Only)
-//! - **E3.1**: Role transition listener triggered on every role change
-//! - **E3.2**: Multiple role transition listeners supported
-//!
-//! ### F. Event Loop Priority and Ordering Tests
-//!
-//! These tests verify the critical event priority ordering in Raft::run() main loop.
-//! The `biased` select! ensures deterministic priority: P0 > P1 > P2 > P3.
-//! This ordering prevents starvation and ensures critical events are processed first.
-//!
-//! Reference: tokio::select! with biased ensures branch order matters
-//!
-//! #### F1. Event Priority Hierarchy
-//!
-//! **Priority Order (P0 > P1 > P2 > P3):**
-//! - **P0 (Shutdown)**: shutdown_signal.changed()
-//!   - **F1.1**: Shutdown always processed first, even if tick/role events pending
-//!   - **F1.2**: Pending events discarded on shutdown
-//!   - **F1.3**: Graceful termination ensures no state corruption
-//!
-//! - **P1 (Tick)**: sleep_until(next_deadline) - election timeout / heartbeat
-//!   - **F1.4**: Tick has highest operational priority
-//!   - **F1.5**: Tick drives election timeout and heartbeat cadence
-//!   - **F1.6**: Tick fires even if role_rx/event_rx have pending messages
-//!
-//! - **P2 (RoleEvent)**: role_rx.recv() - internal state transitions
-//!   - **F1.7**: RoleEvent processed before network events
-//!   - **F1.8**: Leadership/follower transitions take priority over RPCs
-//!   - **F1.9**: Multiple role events processed in order (not all at once)
-//!
-//! - **P3 (RaftEvent)**: event_rx.recv() - network RPCs and responses
-//!   - **F1.10**: Network events (AppendEntries, RequestVote) processed last
-//!   - **F1.11**: Responses from followers/candidates deferred until tick completes
-//!   - **F1.12**: This prevents RPC storms from starving timers
-//!
-//! #### F2. Concurrent Event Arrival Scenarios
-//!
-//! **Scenario: Tick fires while role_event and raft_event pending**
-//! - **F2.1**: Tick processes first (P1 > P2, P3)
-//! - **F2.2**: Role event processes next (P2 > P3)
-//! - **F2.3**: Raft event processes last
-//! - **F2.4**: Starvation prevention: next iteration will service events
-//!
-//! **Scenario: Role event while multiple raft events pending**
-//! - **F2.5**: Single role event dequeued and processed
-//! - **F2.6**: Only one raft event processed per iteration
-//! - **F2.7**: Multiple role events do NOT accumulate in single iteration
-//!
-//! **Scenario: Shutdown signal during election**
-//! - **F2.8**: Shutdown takes priority even if Tick and RoleEvent pending
-//! - **F2.9**: Partial election in progress is abandoned
-//! - **F2.10**: State machine remains consistent (no partial transitions)
-//!
-//! #### F3. Election Timeout Behavior
-//!
-//! These tests verify that the Raft main loop correctly handles election timeouts
-//! using tokio::time::pause() for deterministic timing control.
-//!
-//! - **F3.1**: Follower election timeout → BecomeCandidate
-//! - **F3.2**: Candidate timeout → restart election (stay Candidate or become Follower)
-//! - **F3.3**: Leader timeout → no action (heartbeat continues)
-//! - **F3.4**: Timeout NOT missed when role_event/raft_event arrive
-//! - **F3.5**: Timeout randomization prevents split brain
-//!
-//! #### F4. Complete Election Flow with Event Ordering
-//! - **F4.1**: Follower → Candidate (election starts via tick)
-//! - **F4.2**: Candidate → Leader (majority votes received via raft events)
-//! - **F4.3**: Leadership verified (noop entry replicated via raft events)
-//! - **F4.4**: RoleEvent BecomeLeader processes before any AppendEntries responses
-//! - **F4.5**: Tick continues at regular intervals during replication
-//!
-//! ### G. Error Cases and Edge Cases
-//!
-//! These tests verify robustness and correct error handling.
-//!
-//! #### G1. Invalid State Transitions
-//! - **G1.1**: All invalid transitions return error
-//! - **G1.2**: State unchanged after failed transition
-//!
-//! #### G2. Single-Node Cluster Special Cases
-//! - **G2.1**: Single node is always leader (no peers needed)
-//! - **G2.2**: Election always succeeds (no votes needed)
-//! - **G2.3**: No peer replication needed
-//!
-//! #### G3. Concurrent Event Handling
-//! - **G3.1**: Multiple listeners receiving notifications simultaneously
-//! - **G3.2**: RoleEvents arriving while processing previous event
-//!
-//! #### G4. Listener Registration Edge Cases
-//! - **G4.1**: Register listeners after role transitions
-//! - **G4.2**: Listener registration/deregistration during events
-//!
-//! ### H. Node Joining/Bootstrapping Tests (Cluster Membership Changes)
-//!
-//! These tests cover the critical behavior of new nodes joining an existing cluster.
-//! This is a special initialization path that differs from normal role transitions.
-//! Reference: Raft paper §6 (Cluster membership changes)
-//!
-//! #### H1. New Node Snapshot Initialization
-//! - **H1.1**: is_learner() flag indicates bootstrap phase
-//! - **H1.2**: fetch_initial_snapshot() called in Raft::run() before main loop
-//! - **H1.3**: Snapshot fetch success - node has log prefix from snapshot
-//! - **H1.4**: Snapshot fetch failure - node falls back to append_entries sync
-//! - **H1.5**: Node starts as Follower (not Candidate, no voting yet)
-//! - **H1.6**: After snapshot, ready to participate in normal replication
-//!
-//! #### H2. New Node Replication Catchup
-//! - **H2.1**: New node receives append_entries from Leader
-//! - **H2.2**: Log entries replicated to new node until match_index catches up
-//! - **H2.3**: match_index reaches last_entry_id
-//! - **H2.4**: New node can participate in quorum once caught up
-//! - **H2.5**: Voter vs Learner distinction during bootstrap
-//!
-//! #### H3. Bootstrap Timing and Ordering
-//! - **H3.1**: Snapshot initialization happens BEFORE main event loop starts
-//! - **H3.2**: is_learner() flag checked once per node startup
-//! - **H3.3**: No role transitions or elections during bootstrap phase
-//! - **H3.4**: join_cluster() vs run() method sequencing
-//!
-//! #### H4. Edge Cases in Joining
-//! - **H4.1**: Leader crashes before sending snapshot
-//! - **H4.2**: Multiple new nodes joining simultaneously
-//! - **H4.3**: New node joining with empty log
-//! - **H4.4**: New node joining when cluster has no Leader
-//! - **H4.5**: New node promoted from Learner to Voter during join phase
-//!
-//! #### H5. Consistency Guarantees During Join
-//! - **H5.1**: New node cannot commit entries before catch-up
-//! - **H5.2**: New node cannot vote before catch-up (if learner)
-//! - **H5.3**: Snapshot consistency: snapshot term/index must be valid
-//! - **H5.4**: Log prefix matching: new entries follow snapshot boundary
-//!
-//! ## Test Infrastructure
-//!
-//! ### Tools and Utilities
-//! - `MockBuilder`: Create minimal Raft instances with mocked dependencies
-//! - `MockMembership`: Mock cluster topology
-//! - `MockRaftLog`: Mock log storage
-//! - `MockReplicationCore`: Mock log replication
-//! - `MockElectionCore`: Mock voting
-//! - `tokio::time::pause()` + `advance()`: Deterministic timing control
-//! - `mpsc/watch` channels: Listener notification testing
-//!
-//! ### Raft Protocol Compliance
-//! All tests must follow the Raft paper precisely:
-//! - Raft paper: https://raft.github.io
-//! - Leader election rules (§5.2)
-//! - Log replication rules (§5.3)
-//! - Quorum and safety properties (§5.4)
-//! - Leadership verification (§5.4.2)
-//!
-//! ### Test Characteristics
-//! - Deterministic: Use fake timers for time-dependent tests
-//! - Isolated: Mock all external dependencies
-//! - Fast: Complete suite should run in <10 seconds
-//! - Comprehensive: Cover normal path, error path, and edge cases
-//! - Clear: Test names and comments explain what is being tested and why
 
-#[cfg(test)]
-mod leader_change_tests {
-    use tokio::sync::mpsc;
-
-    #[test]
-    fn test_leader_change_listener_registration() {
-        // Test that we can create channels for leader change notifications
-        let (tx, mut rx) = mpsc::unbounded_channel::<(Option<u32>, u64)>();
-
-        // Simulate sending a notification
-        tx.send((Some(1), 5)).unwrap();
-
-        // Verify we can receive it
-        let (leader_id, term) = rx.try_recv().expect("Should receive notification");
-        assert_eq!(leader_id, Some(1));
-        assert_eq!(term, 5);
-    }
-
-    #[test]
-    fn test_multiple_listeners() {
-        // Test broadcasting to multiple listeners
-        let (tx1, mut rx1) = mpsc::unbounded_channel::<(Option<u32>, u64)>();
-        let (tx2, mut rx2) = mpsc::unbounded_channel::<(Option<u32>, u64)>();
-
-        // Simulate sending to both
-        tx1.send((Some(2), 10)).unwrap();
-        tx2.send((Some(2), 10)).unwrap();
-
-        // Verify both receive
-        let (leader1, term1) = rx1.try_recv().expect("Listener 1 should receive");
-        let (leader2, term2) = rx2.try_recv().expect("Listener 2 should receive");
-
-        assert_eq!(leader1, Some(2));
-        assert_eq!(term1, 10);
-        assert_eq!(leader2, Some(2));
-        assert_eq!(term2, 10);
-    }
-
-    #[test]
-    fn test_no_leader_notification() {
-        // Test sending None for leader_id (candidate state)
-        let (tx, mut rx) = mpsc::unbounded_channel::<(Option<u32>, u64)>();
-
-        tx.send((None, 15)).unwrap();
-
-        let (leader_id, term) = rx.try_recv().expect("Should receive notification");
-        assert_eq!(leader_id, None);
-        assert_eq!(term, 15);
-    }
-
-    #[test]
-    fn test_channel_closed() {
-        // Test that sending fails when receiver is dropped
-        let (tx, rx) = mpsc::unbounded_channel::<(Option<u32>, u64)>();
-
-        drop(rx);
-
-        let result = tx.send((Some(1), 5));
-        assert!(result.is_err(), "Send should fail when receiver is dropped");
-    }
-}
-
-#[cfg(test)]
-mod leader_discovered_tests {
-    use tokio::sync::watch;
-
-    use super::super::Raft;
-    use super::super::RoleEvent;
-    use crate::test_utils::MockBuilder;
-    use crate::test_utils::MockTypeConfig;
-
-    #[tokio::test]
-    async fn test_leader_discovered_event_handling() {
-        // Test that LeaderDiscovered event triggers leader change notification
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft: Raft<MockTypeConfig> = MockBuilder::new(graceful_rx).build_raft();
-
-        // Register leader change listener
-        let (leader_tx, mut leader_rx) = watch::channel(None);
-        raft.register_leader_change_listener(leader_tx);
-
-        // Send LeaderDiscovered event
-        let leader_id = 3;
-        let term = 5;
-        raft.handle_role_event(RoleEvent::LeaderDiscovered(leader_id, term))
-            .await
-            .expect("Should handle LeaderDiscovered");
-
-        // Verify notification was sent
-        leader_rx.changed().await.expect("Should receive change notification");
-        let leader_info = *leader_rx.borrow();
-        assert!(leader_info.is_some());
-        let info = leader_info.unwrap();
-        assert_eq!(info.leader_id, leader_id);
-        assert_eq!(info.term, term);
-    }
-
-    #[tokio::test]
-    async fn test_leader_discovered_no_state_change() {
-        // Test that LeaderDiscovered does NOT change node role
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft: Raft<MockTypeConfig> = MockBuilder::new(graceful_rx).build_raft();
-
-        let initial_role = raft.role.as_i32();
-
-        // Send LeaderDiscovered event
-        raft.handle_role_event(RoleEvent::LeaderDiscovered(3, 5))
-            .await
-            .expect("Should handle LeaderDiscovered");
-
-        // Verify role unchanged (still Follower)
-        assert_eq!(raft.role.as_i32(), initial_role);
-    }
-
-    #[tokio::test]
-    async fn test_leader_discovered_multiple_listeners() {
-        // Test that multiple subscribers can receive notifications via watch::Sender::subscribe()
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft: Raft<MockTypeConfig> = MockBuilder::new(graceful_rx).build_raft();
-
-        // Register leader change listener
-        let (tx, _rx) = watch::channel(None);
-        raft.register_leader_change_listener(tx.clone());
-
-        // Create multiple subscribers
-        let mut rx1 = tx.subscribe();
-        let mut rx2 = tx.subscribe();
-
-        // Send LeaderDiscovered event
-        let leader_id = 2;
-        let term = 10;
-        raft.handle_role_event(RoleEvent::LeaderDiscovered(leader_id, term))
-            .await
-            .expect("Should handle LeaderDiscovered");
-
-        // Verify all subscribers receive notification
-        rx1.changed().await.expect("Subscriber 1 should receive");
-        rx2.changed().await.expect("Subscriber 2 should receive");
-
-        let info1 = (*rx1.borrow()).unwrap();
-        let info2 = (*rx2.borrow()).unwrap();
-
-        assert_eq!(info1.leader_id, leader_id);
-        assert_eq!(info1.term, term);
-        assert_eq!(info2.leader_id, leader_id);
-        assert_eq!(info2.term, term);
-    }
-
-    #[tokio::test]
-    async fn test_leader_discovered_with_deduplication() {
-        // Test that watch channel automatically deduplicates identical notifications
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft: Raft<MockTypeConfig> = MockBuilder::new(graceful_rx).build_raft();
-
-        let (leader_tx, mut leader_rx) = watch::channel(None);
-        raft.register_leader_change_listener(leader_tx);
-
-        // Send same leader multiple times
-        raft.handle_role_event(RoleEvent::LeaderDiscovered(2, 5))
-            .await
-            .expect("Should handle first");
-        raft.handle_role_event(RoleEvent::LeaderDiscovered(2, 5))
-            .await
-            .expect("Should handle second (duplicate)");
-
-        // Should receive only one notification (watch channel deduplicates)
-        leader_rx.changed().await.expect("Should receive first change");
-        let info = (*leader_rx.borrow()).unwrap();
-        assert_eq!(info.leader_id, 2);
-        assert_eq!(info.term, 5);
-
-        // No second change notification because value is identical
-        tokio::select! {
-            _ = leader_rx.changed() => {
-                panic!("Should not receive duplicate notification");
-            }
-            _ = tokio::time::sleep(tokio::time::Duration::from_millis(100)) => {
-                // Expected: timeout because no new change
-            }
-        }
-    }
-
-    #[test]
-    fn test_role_event_leader_discovered_creation() {
-        // Test creating LeaderDiscovered event
-        let leader_id = 5;
-        let term = 20;
-        let event = RoleEvent::LeaderDiscovered(leader_id, term);
-
-        // Verify we can match on it
-        match event {
-            RoleEvent::LeaderDiscovered(id, t) => {
-                assert_eq!(id, leader_id);
-                assert_eq!(t, term);
-            }
-            _ => panic!("Should be LeaderDiscovered variant"),
-        }
-    }
-}
-
-#[cfg(test)]
 mod raft_comprehensive_tests {
     use std::sync::Arc;
 
@@ -521,8 +29,8 @@ mod raft_comprehensive_tests {
     use tokio::sync::mpsc;
     use tokio::sync::watch;
 
-    use super::super::NewCommitData;
-    use super::super::RoleEvent;
+    use crate::NewCommitData;
+    use crate::RoleEvent;
     use crate::test_utils::MockBuilder;
     use crate::test_utils::MockTypeConfig;
 
@@ -2527,1218 +2035,3 @@ mod raft_comprehensive_tests {
     }
 }
 
-// ============================================================================
-// P0: CRITICAL PATH TESTS - DRAIN-BASED BATCH ARCHITECTURE
-// ============================================================================
-//
-// These tests verify the core drain-based batch processing mechanism introduced
-// in #266 refactor. The drain pattern (recv → try_recv drains → flush) replaces
-// timeout-driven batching to eliminate low-latency overhead while improving
-// high-load throughput.
-//
-#[cfg(test)]
-mod drain_based_batch_architecture_tests {
-    use crate::RaftOneshot;
-    use crate::maybe_clone_oneshot::MaybeCloneOneshot;
-    use crate::{
-        MockBuilder,
-        raft_role::{RaftRole, role_state::RaftRoleState},
-        replication::{BatchBuffer, BatchMetrics, BatchTriggerType},
-    };
-    use bytes::Bytes;
-    use d_engine_proto::client::write_command::Operation;
-    use d_engine_proto::client::{
-        ClientReadRequest, ClientWriteRequest, ReadConsistencyPolicy, WriteCommand,
-    };
-    use std::sync::Arc;
-    use std::time::Duration;
-    use tokio::sync::watch;
-    use tokio::time::Instant;
-
-    // ========================================================================
-    // P0-1: LOW LOAD - SINGLE COMMAND (ZERO LATENCY)
-    // ========================================================================
-    //
-    // **Objective**: Verify batch=1 with no unnecessary wait in low load
-    //
-    // **Expected Behavior**:
-    // - recv() blocks for first command → try_recv() finds empty channel → immediate flush
-    // - batch.size = 1, no timeout penalty
-    // - Latency < 5ms (vs old architecture ~10ms with batch_timeout)
-    //
-    #[test]
-    fn test_p0_batch_buffer_single_item_no_batching() {
-        // **Setup**: Create BatchBuffer with small max_batch_size
-        let metrics = Arc::new(BatchMetrics::new(1));
-        let mut buffer: BatchBuffer<u32> = BatchBuffer::new(300).with_metrics(metrics.clone());
-
-        let start = Instant::now();
-
-        // **Step 1**: Push single item
-        buffer.push(42);
-        assert_eq!(buffer.len(), 1);
-
-        // **Step 2**: Drain (take_all) without timeout
-        let drained = buffer.take_with_trigger(BatchTriggerType::Drain);
-        let elapsed = start.elapsed();
-
-        // **Assertions**:
-        // ✅ Single item drained immediately
-        assert_eq!(drained.len(), 1);
-        assert_eq!(drained[0], 42);
-
-        // ✅ Buffer is now empty
-        assert!(buffer.is_empty());
-
-        // ✅ Latency is minimal (no batching wait)
-        assert!(
-            elapsed < Duration::from_millis(5),
-            "Drain latency should be <5ms, got {:?}",
-            elapsed
-        );
-
-        // ✅ Metrics: drain_triggered = 1, batch_size = 1
-        assert_eq!(
-            metrics.drain_triggered.load(std::sync::atomic::Ordering::Relaxed),
-            1
-        );
-        assert_eq!(
-            metrics.total_batch_size.load(std::sync::atomic::Ordering::Relaxed),
-            1
-        );
-    }
-
-    // ========================================================================
-    // P0-2: HIGH LOAD - MAX BATCH CAP (PREVENT OVERLOAD)
-    // ========================================================================
-    //
-    // **Objective**: Verify `max_batch_size` limits single drain to prevent IO overload
-    //
-    // **Expected Behavior**:
-    // - First drain: reaches max_batch_size (300), stops draining
-    // - Multiple cycles until all items processed
-    // - Throughput > 300k ops/sec (vs old ~72k ops/sec, 4x improvement)
-    //
-    #[test]
-    fn test_p0_batch_buffer_max_batch_size_cap() {
-        // **Setup**: Create BatchBuffer with max_batch_size = 300
-        let metrics = Arc::new(BatchMetrics::new(1));
-        let mut buffer: BatchBuffer<u32> = BatchBuffer::new(300).with_metrics(metrics.clone());
-
-        const TOTAL_ITEMS: usize = 1000;
-
-        // **Step 1**: Push 1000 items (simulating high-speed client)
-        for i in 0..TOTAL_ITEMS {
-            buffer.push(i as u32);
-        }
-        assert_eq!(buffer.len(), TOTAL_ITEMS);
-
-        // **Step 2**: First drain cycle - should cap at max_batch_size
-        let batch1 = buffer.take_with_trigger(BatchTriggerType::Drain);
-        assert_eq!(
-            batch1.len(),
-            300,
-            "First drain should respect max_batch_size"
-        );
-        assert_eq!(buffer.len(), TOTAL_ITEMS - 300);
-
-        // **Step 3**: Continue draining - should process remaining in multiple cycles
-        let mut total_drained = 300;
-        let mut drain_cycles = 1;
-
-        while !buffer.is_empty() {
-            let batch = buffer.take_with_trigger(BatchTriggerType::Drain);
-            assert!(
-                batch.len() <= 300,
-                "Batch size should never exceed max_batch_size"
-            );
-            total_drained += batch.len();
-            drain_cycles += 1;
-        }
-
-        // **Assertions**:
-        // ✅ All items eventually drained
-        assert_eq!(total_drained, TOTAL_ITEMS);
-
-        // ✅ Multiple drain cycles (not single batch)
-        assert!(
-            drain_cycles >= 3,
-            "Should have multiple drain cycles for 1000 items with batch_size=300"
-        );
-
-        // ✅ Each batch ≤ max_batch_size (IO overload prevention)
-        // (verified in step 3 loop)
-
-        // ✅ Metrics: drain_triggered = number of cycles
-        assert_eq!(
-            metrics.drain_triggered.load(std::sync::atomic::Ordering::Relaxed),
-            drain_cycles as u64
-        );
-    }
-
-    // ========================================================================
-    // P0-3: BATCH BUFFER - EMPTY BUFFER IDEMPOTENCY
-    // ========================================================================
-    //
-    // **Objective**: Verify empty buffer flush is safe (no panic, no-op)
-    //
-    // **Expected Behavior**:
-    // - is_empty() returns true → flush is no-op
-    // - No error, no panic
-    // - System remains stable
-    //
-    #[test]
-    fn test_p0_batch_buffer_empty_flush_idempotent() {
-        // **Setup**: Create BatchBuffer
-        let metrics = Arc::new(BatchMetrics::new(1));
-        let mut buffer: BatchBuffer<u32> = BatchBuffer::new(300).with_metrics(metrics.clone());
-
-        // **Step 1**: Push and drain one item
-        buffer.push(42);
-        let drained1 = buffer.take_with_trigger(BatchTriggerType::Drain);
-        assert_eq!(drained1.len(), 1);
-
-        // **Step 2**: Verify buffer is empty
-        assert!(buffer.is_empty());
-
-        // **Step 3**: Try to drain empty buffer (should be no-op)
-        let drained2 = buffer.take_with_trigger(BatchTriggerType::Drain);
-
-        // **Assertions**:
-        // ✅ is_empty() returns true
-        assert!(buffer.is_empty());
-
-        // ✅ Second drain returns empty VecDeque
-        assert_eq!(drained2.len(), 0);
-
-        // ✅ Metrics recorded both flushes (even though second is empty)
-        assert_eq!(
-            metrics.drain_triggered.load(std::sync::atomic::Ordering::Relaxed),
-            2
-        );
-        assert_eq!(
-            metrics.total_batch_size.load(std::sync::atomic::Ordering::Relaxed),
-            1
-        );
-
-        // ✅ No panic, system stable
-    }
-
-    // ========================================================================
-    // P0-4: BATCH BUFFER - MAX BATCH SIZE = 1 (DEGENERATE CASE)
-    // ========================================================================
-    //
-    // **Objective**: Verify system works when max_batch_size=1 (disable batching)
-    //
-    // **Expected Behavior**:
-    // - Each drain processes exactly 1 command
-    // - Multiple drain cycles for N items
-    // - Still better latency than old timeout-driven architecture
-    //
-    #[test]
-    fn test_p0_batch_buffer_max_batch_size_one() {
-        // **Setup**: Create BatchBuffer with max_batch_size = 1 (degenerate)
-        let metrics = Arc::new(BatchMetrics::new(1));
-        let mut buffer: BatchBuffer<u32> = BatchBuffer::new(1).with_metrics(metrics.clone());
-
-        const ITEMS: usize = 10;
-
-        // **Step 1**: Push 10 items
-        for i in 0..ITEMS {
-            buffer.push(i as u32);
-        }
-
-        // **Step 2**: Drain one by one
-        let mut drain_count = 0;
-        while !buffer.is_empty() {
-            let batch = buffer.take_with_trigger(BatchTriggerType::Drain);
-            assert_eq!(batch.len(), 1, "Each drain should process exactly 1 item");
-            drain_count += 1;
-        }
-
-        // **Assertions**:
-        // ✅ 10 items = 10 drain cycles
-        assert_eq!(drain_count, ITEMS);
-
-        // ✅ All items processed
-        assert!(buffer.is_empty());
-
-        // ✅ Graceful degradation (no panic, system works)
-        assert_eq!(
-            metrics.drain_triggered.load(std::sync::atomic::Ordering::Relaxed),
-            ITEMS as u64
-        );
-    }
-
-    // ========================================================================
-    // P0-5: BATCH METRICS - TRIGGER TYPE TRACKING
-    // ========================================================================
-    //
-    // **Objective**: Verify metrics correctly track drain vs heartbeat triggers
-    //
-    // **Expected Behavior**:
-    // - drain_triggered counts Drain triggers
-    // - heartbeat_triggered counts Heartbeat triggers
-    // - total_batch_count = drain + heartbeat
-    // - total_batch_size tracks sum of all batch sizes
-    //
-    #[test]
-    fn test_p0_batch_metrics_drain_vs_heartbeat_tracking() {
-        // **Setup**: Create BatchMetrics
-        let metrics = Arc::new(BatchMetrics::new(1));
-        let mut buffer: BatchBuffer<u32> = BatchBuffer::new(300).with_metrics(metrics.clone());
-
-        // **Step 1**: Push items and trigger drain flushes
-        for i in 0..100 {
-            buffer.push(i as u32);
-        }
-
-        let batch1 = buffer.take_with_trigger(BatchTriggerType::Drain);
-        assert_eq!(batch1.len(), 100);
-
-        // **Step 2**: Push more items and trigger heartbeat flush
-        for i in 100..150 {
-            buffer.push(i as u32);
-        }
-
-        let batch2 = buffer.take_with_trigger(BatchTriggerType::Heartbeat);
-        assert_eq!(batch2.len(), 50);
-
-        // **Step 3**: One more drain
-        for i in 150..200 {
-            buffer.push(i as u32);
-        }
-
-        let batch3 = buffer.take_with_trigger(BatchTriggerType::Drain);
-        assert_eq!(batch3.len(), 50);
-
-        // **Assertions**:
-        // ✅ drain_triggered = 2 (batch1 and batch3)
-        assert_eq!(
-            metrics.drain_triggered.load(std::sync::atomic::Ordering::Relaxed),
-            2,
-            "Should have 2 Drain triggers"
-        );
-
-        // ✅ heartbeat_triggered = 1 (batch2)
-        assert_eq!(
-            metrics.heartbeat_triggered.load(std::sync::atomic::Ordering::Relaxed),
-            1,
-            "Should have 1 Heartbeat trigger"
-        );
-
-        // ✅ total_batch_count = 3 (all flushes)
-        assert_eq!(
-            metrics.total_batch_count.load(std::sync::atomic::Ordering::Relaxed),
-            3,
-            "Should have 3 total batch flushes"
-        );
-
-        // ✅ total_batch_size = 200 (100 + 50 + 50)
-        assert_eq!(
-            metrics.total_batch_size.load(std::sync::atomic::Ordering::Relaxed),
-            200,
-            "Total batch size should be 200"
-        );
-    }
-
-    // ========================================================================
-    // P0-6: MEDIUM LOAD - NATURAL BATCHING
-    // ========================================================================
-    //
-    // **Objective**: Verify batching respects max_batch_size under moderate load
-    //
-    // **Expected Behavior**:
-    // - First drain: limited by max_batch_size (20 items)
-    // - Multiple drain cycles (not single batch) when total > max_batch_size
-    // - Each batch ≤ max_batch_size
-    //
-    #[test]
-    fn test_p0_batch_buffer_natural_batching_medium_load() {
-        // **Setup**: Create BatchBuffer with small max_batch_size to demonstrate batching
-        let metrics = Arc::new(BatchMetrics::new(1));
-        let mut buffer: BatchBuffer<u32> = BatchBuffer::new(20).with_metrics(metrics.clone());
-
-        // **Step 1**: Simulate moderate load - 50 items
-        // With max_batch_size=20, should require multiple drain cycles
-        for i in 0..50 {
-            buffer.push(i as u32);
-        }
-
-        // **Step 2**: First drain - should be capped at max_batch_size
-        let batch1 = buffer.take_with_trigger(BatchTriggerType::Drain);
-        let first_batch_size = batch1.len();
-
-        // First batch should respect max_batch_size limit
-        assert_eq!(
-            first_batch_size, 20,
-            "First batch should be capped at max_batch_size=20"
-        );
-
-        // **Step 3**: Drain remaining items
-        let mut total_drained = first_batch_size;
-        let mut drain_cycles = 1;
-
-        while !buffer.is_empty() {
-            let batch = buffer.take_with_trigger(BatchTriggerType::Drain);
-            assert!(
-                batch.len() <= 20,
-                "Each batch should respect max_batch_size"
-            );
-            total_drained += batch.len();
-            drain_cycles += 1;
-        }
-
-        // **Assertions**:
-        // ✅ All items drained
-        assert_eq!(total_drained, 50);
-
-        // ✅ Multiple cycles (with 50 items and max_batch_size=20, expect 3 cycles)
-        assert!(
-            drain_cycles > 1,
-            "Should have multiple drain cycles for natural batching"
-        );
-
-        // ✅ Exactly 3 cycles: 20 + 20 + 10
-        assert_eq!(drain_cycles, 3, "Expected 3 drain cycles: 20 + 20 + 10");
-    }
-
-    // ========================================================================
-    // P0 UNIT TESTS - ROLE-SPECIFIC DRAIN BEHAVIOR
-    // ========================================================================
-    //
-    // These tests verify the drain-based batch architecture at the RoleState level.
-    // Each role (Leader/Follower/Candidate/Learner) has specific behavior for
-    // push_client_cmd() and flush_cmd_buffers().
-    //
-    // Reference: 001_todo_raft_test_plan_2215.md P0-1 through P0-5
-    //
-
-    // ========================================================================
-    // P0-1: LEADER - SINGLE COMMAND DRAIN IMMEDIATELY
-    // ========================================================================
-    //
-    // **Objective**: Verify Leader processes single command with batch=1, no delay
-    //
-    // **Expected Behavior**:
-    // - Command enters propose_buffer via push_client_cmd()
-    // - flush_cmd_buffers() drains immediately
-    // - Metrics: drain_triggered=1, batch_size=1
-    //
-    #[tokio::test]
-    async fn test_p0_leader_single_command_drain_immediately() {
-        // **Setup**: Create Leader Raft instance
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup mocks
-        let mut raft_log = crate::MockRaftLog::new();
-        raft_log.expect_last_entry_id().returning(|| 11);
-        raft_log.expect_flush().returning(|| Ok(()));
-        raft_log.expect_append_entries().returning(|_| Ok(()));
-        raft_log.expect_calculate_majority_matched_index().returning(|_, _, _| Some(11));
-        raft_log.expect_load_hard_state().returning(|| Ok(None));
-        raft_log.expect_save_hard_state().returning(|_| Ok(()));
-
-        let mut replication_handler = crate::MockReplicationCore::new();
-        replication_handler
-            .expect_handle_raft_request_in_batch()
-            .returning(|_, _, _, _, _| {
-                Ok(crate::AppendResults {
-                    commit_quorum_achieved: true,
-                    learner_progress: std::collections::HashMap::new(),
-                    peer_updates: std::collections::HashMap::new(),
-                })
-            });
-
-        raft.ctx.storage.raft_log = Arc::new(raft_log);
-        raft.ctx.handlers.replication_handler = replication_handler;
-
-        // Transition to Leader
-        raft.handle_role_event(crate::RoleEvent::BecomeCandidate)
-            .await
-            .expect("Should become Candidate");
-        raft.handle_role_event(crate::RoleEvent::BecomeLeader)
-            .await
-            .expect("Should become Leader");
-
-        // **Step 1**: Send single write command
-
-        let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-        let write_cmd = WriteCommand {
-            operation: Some(Operation::Insert(
-                d_engine_proto::client::write_command::Insert {
-                    key: Bytes::from("test_key"),
-                    value: Bytes::from("test_value"),
-                    ttl_secs: 0,
-                },
-            )),
-        };
-        let write_req = ClientWriteRequest {
-            client_id: 1,
-            commands: vec![write_cmd],
-        };
-
-        // Access leader state to push command
-        if let RaftRole::Leader(ref mut leader) = raft.role {
-            let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-
-            // **Step 2**: Push command to buffer
-            leader.push_client_cmd(cmd, &raft.ctx);
-
-            // **Step 3**: Flush buffers (simulate drain)
-            let (role_tx, _role_rx) = tokio::sync::mpsc::unbounded_channel();
-            leader
-                .flush_cmd_buffers(&raft.ctx, &role_tx)
-                .await
-                .expect("flush should succeed");
-
-            // **Assertion**: flush_cmd_buffers completed successfully
-            // (Private buffer fields are implementation details - not tested directly)
-        } else {
-            panic!("Expected Leader state");
-        }
-    }
-
-    // ========================================================================
-    // P0-2: LEADER - HIGH LOAD MAX BATCH CAP
-    // ========================================================================
-    //
-    // **Objective**: Verify Leader respects max_batch_size=300 under high load
-    //
-    // **Expected Behavior**:
-    // - Push 1000 commands to propose_buffer
-    // - Each flush processes ≤ 300 commands
-    // - Multiple drain cycles until all processed
-    //
-    #[tokio::test]
-    async fn test_p0_leader_high_load_max_batch_cap() {
-        // **Setup**: Create Leader with max_batch_size monitoring
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup mocks - allow multiple flush calls
-        let mut raft_log = crate::MockRaftLog::new();
-        raft_log.expect_last_entry_id().returning(|| 11);
-        raft_log.expect_flush().returning(|| Ok(()));
-        raft_log.expect_append_entries().returning(|_| Ok(()));
-        raft_log.expect_calculate_majority_matched_index().returning(|_, _, _| Some(11));
-        raft_log.expect_load_hard_state().returning(|| Ok(None));
-        raft_log.expect_save_hard_state().returning(|_| Ok(()));
-
-        let mut replication_handler = crate::MockReplicationCore::new();
-        replication_handler
-            .expect_handle_raft_request_in_batch()
-            .returning(|_, _, _, _, _| {
-                Ok(crate::AppendResults {
-                    commit_quorum_achieved: true,
-                    learner_progress: std::collections::HashMap::new(),
-                    peer_updates: std::collections::HashMap::new(),
-                })
-            });
-
-        raft.ctx.storage.raft_log = Arc::new(raft_log);
-        raft.ctx.handlers.replication_handler = replication_handler;
-
-        // Transition to Leader
-        raft.handle_role_event(crate::RoleEvent::BecomeCandidate)
-            .await
-            .expect("Should become Candidate");
-        raft.handle_role_event(crate::RoleEvent::BecomeLeader)
-            .await
-            .expect("Should become Leader");
-
-        // **Step 1**: Push 1000 commands rapidly (simulate high load)
-
-        const TOTAL_COMMANDS: usize = 1000;
-
-        if let RaftRole::Leader(ref mut leader) = raft.role {
-            for i in 0..TOTAL_COMMANDS {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let write_cmd = WriteCommand {
-                    operation: Some(Operation::Insert(
-                        d_engine_proto::client::write_command::Insert {
-                            key: Bytes::from(format!("key_{}", i)),
-                            value: Bytes::from(format!("value_{}", i)),
-                            ttl_secs: 0,
-                        },
-                    )),
-                };
-                let write_req = ClientWriteRequest {
-                    client_id: 1,
-                    commands: vec![write_cmd],
-                };
-                let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-            }
-
-            // **Step 2**: Flush to drain
-            let (role_tx, _role_rx) = tokio::sync::mpsc::unbounded_channel();
-            leader
-                .flush_cmd_buffers(&raft.ctx, &role_tx)
-                .await
-                .expect("flush should succeed");
-        } else {
-            panic!("Expected Leader state");
-        }
-    }
-
-    // ========================================================================
-    // P0-3: FOLLOWER - EVENTUAL READ LOCAL PROCESSING
-    // ========================================================================
-    //
-    // **Objective**: Verify Follower processes Eventual reads locally (no redirect)
-    //
-    // **Expected Behavior**:
-    // - Eventual read processed immediately in push_client_cmd()
-    // - No queue buffering
-    // - Reads served from local state machine
-    // - No NOT_LEADER errors
-    //
-    #[tokio::test]
-    async fn test_p0_follower_eventual_read_local_processing() {
-        // **Setup**: Create Follower Raft instance
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup state machine mock for read
-        let mut state_machine_handler = crate::MockStateMachineHandler::new();
-        state_machine_handler
-            .expect_read_from_state_machine()
-            .times(1)
-            .withf(|keys| keys.len() == 1 && keys[0] == bytes::Bytes::from("test_key"))
-            .returning(|_| {
-                Some(vec![d_engine_proto::client::ClientResult {
-                    key: bytes::Bytes::from("test_key"),
-                    value: bytes::Bytes::from("test_value"),
-                }])
-            });
-        raft.ctx.handlers.state_machine_handler = Arc::new(state_machine_handler);
-
-        // Verify initial state is Follower
-        assert!(matches!(raft.role, RaftRole::Follower(_)));
-
-        // **Step 1**: Send Eventual read request to Follower
-        let (response_tx, mut response_rx) = MaybeCloneOneshot::new();
-        let read_req = ClientReadRequest {
-            client_id: 1,
-            keys: vec![bytes::Bytes::from("test_key")],
-            consistency_policy: Some(ReadConsistencyPolicy::EventualConsistency as i32),
-        };
-
-        // Access follower state to push read command
-        if let RaftRole::Follower(ref mut follower) = raft.role {
-            let cmd = crate::ClientCmd::Read(read_req, response_tx);
-
-            // **Step 2**: Push read command (should process immediately)
-            follower.push_client_cmd(cmd, &raft.ctx);
-
-            // **Assertion 1**: Response should be ready immediately
-            // Eventual reads are processed synchronously in push_client_cmd
-            let result = response_rx.recv().await;
-            assert!(result.is_ok(), "Eventual read should return response");
-
-            // **Assertion 2**: Response should be success (not NOT_LEADER error)
-            if let Ok(response) = result {
-                match response {
-                    Ok(read_response) => {
-                        // Check success_result for ReadResults
-                        assert!(
-                            read_response.success_result.is_some(),
-                            "Should have success_result"
-                        );
-                        if let Some(
-                            d_engine_proto::client::client_response::SuccessResult::ReadData(
-                                read_data,
-                            ),
-                        ) = read_response.success_result
-                        {
-                            assert!(!read_data.results.is_empty(), "Should have read results");
-                            assert_eq!(
-                                read_data.results[0].value,
-                                bytes::Bytes::from("test_value")
-                            );
-                        }
-                    }
-                    Err(e) => {
-                        panic!(
-                            "Eventual read should succeed on Follower, got error: {:?}",
-                            e
-                        );
-                    }
-                }
-            }
-        } else {
-            panic!("Expected Follower state");
-        }
-    }
-
-    // ========================================================================
-    // P0-4: FOLLOWER - WRITE REQUEST REJECTION
-    // ========================================================================
-    //
-    // **Objective**: Verify Follower rejects write requests with NOT_LEADER error
-    //
-    // **Expected Behavior**:
-    // - Write request immediately rejected in push_client_cmd()
-    // - Error: NOT_LEADER with Leader information
-    // - No buffer entry
-    // - Response time < 1ms (immediate rejection)
-    //
-    #[tokio::test]
-    async fn test_p0_follower_write_request_rejection() {
-        // **Setup**: Create Follower Raft instance with known leader
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // **Step 1**: Send write request to Follower
-
-        let (response_tx, mut response_rx) = MaybeCloneOneshot::new();
-        let write_cmd = WriteCommand {
-            operation: Some(Operation::Insert(
-                d_engine_proto::client::write_command::Insert {
-                    key: Bytes::from("test_key"),
-                    value: Bytes::from("test_value"),
-                    ttl_secs: 0,
-                },
-            )),
-        };
-        let write_req = ClientWriteRequest {
-            client_id: 1,
-            commands: vec![write_cmd],
-        };
-
-        let start = tokio::time::Instant::now();
-
-        if let RaftRole::Follower(ref mut follower) = raft.role {
-            let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-
-            // **Step 2**: Push write command (should reject immediately)
-            follower.push_client_cmd(cmd, &raft.ctx);
-
-            // **Assertion 1**: Response should be ready immediately
-            let result = response_rx.recv().await;
-            let elapsed = start.elapsed();
-
-            assert!(result.is_ok(), "Should receive response from Follower");
-
-            // **Assertion 2**: Response time < 1ms (immediate rejection)
-            assert!(
-                elapsed.as_millis() < 10,
-                "Write rejection should be immediate, took {:?}ms",
-                elapsed.as_millis()
-            );
-
-            // **Assertion 3**: Response should be NOT_LEADER error
-            if let Ok(response) = result {
-                match response {
-                    Err(err) => {
-                        // Verify it's a NOT_LEADER error (FailedPrecondition or similar)
-                        let err_str = format!("{:?}", err);
-                        assert!(
-                            err_str.contains("Not leader")
-                                || err_str.contains("NotLeader")
-                                || err_str.contains("NOT_LEADER")
-                                || err_str.contains("FailedPrecondition"),
-                            "Expected NOT_LEADER error, got: {:?}",
-                            err
-                        );
-                    }
-                    Ok(_) => {
-                        panic!("Write to Follower should return NOT_LEADER error, got success");
-                    }
-                }
-            }
-        } else {
-            panic!("Expected Follower state");
-        }
-    }
-
-    // ========================================================================
-    // P0-5: CANDIDATE - ALL CLIENT COMMANDS HANDLING
-    // ========================================================================
-    //
-    // **Objective**: Verify Candidate role behavior for different command types
-    //
-    // **Expected Behavior**:
-    // - Write: NOT_LEADER error (immediate rejection)
-    // - Linear read: NOT_LEADER error (immediate rejection)
-    // - Eventual read: SUCCESS (served locally) ← New behavior
-    // - All responses < 2ms
-    //
-    #[tokio::test]
-    async fn test_p0_candidate_commands_handling() {
-        // **Setup**: Create Candidate Raft instance
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup state machine for eventual read
-        let mut state_machine_handler = crate::MockStateMachineHandler::new();
-        state_machine_handler
-            .expect_read_from_state_machine()
-            .times(1)
-            .withf(|keys| keys.len() == 1 && keys[0] == bytes::Bytes::from("eventual_key"))
-            .returning(|_| {
-                Some(vec![d_engine_proto::client::ClientResult {
-                    key: bytes::Bytes::from("eventual_key"),
-                    value: bytes::Bytes::from("eventual_value"),
-                }])
-            });
-        raft.ctx.handlers.state_machine_handler = Arc::new(state_machine_handler);
-
-        // Transition to Candidate
-        raft.handle_role_event(crate::RoleEvent::BecomeCandidate)
-            .await
-            .expect("Should become Candidate");
-
-        // **Test 1**: Write request should be rejected
-        {
-            let (response_tx, mut response_rx) = MaybeCloneOneshot::new();
-            let write_cmd = WriteCommand {
-                operation: Some(Operation::Insert(
-                    d_engine_proto::client::write_command::Insert {
-                        key: Bytes::from("write_key"),
-                        value: Bytes::from("write_value"),
-                        ttl_secs: 0,
-                    },
-                )),
-            };
-            let write_req = ClientWriteRequest {
-                client_id: 1,
-                commands: vec![write_cmd],
-            };
-
-            if let RaftRole::Candidate(ref mut candidate) = raft.role {
-                let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-                candidate.push_client_cmd(cmd, &raft.ctx);
-
-                let result = response_rx.recv().await;
-                assert!(result.is_ok(), "Should receive response");
-
-                // **Assertion 1**: Write rejected with NOT_LEADER
-                if let Ok(Err(err)) = result {
-                    let err_str = format!("{:?}", err);
-                    assert!(
-                        err_str.contains("Not leader")
-                            || err_str.contains("NotLeader")
-                            || err_str.contains("NOT_LEADER")
-                            || err_str.contains("FailedPrecondition"),
-                        "Expected NOT_LEADER for write, got: {:?}",
-                        err
-                    );
-                } else {
-                    panic!("Write to Candidate should return NOT_LEADER error");
-                }
-            } else {
-                panic!("Expected Candidate state");
-            }
-        }
-
-        // **Test 2**: Linear read should be rejected
-        {
-            let (response_tx, mut response_rx) = MaybeCloneOneshot::new();
-            let read_req = ClientReadRequest {
-                client_id: 1,
-                keys: vec![bytes::Bytes::from("linear_key")],
-                consistency_policy: Some(ReadConsistencyPolicy::LinearizableRead as i32),
-            };
-
-            if let RaftRole::Candidate(ref mut candidate) = raft.role {
-                let cmd = crate::ClientCmd::Read(read_req, response_tx);
-                candidate.push_client_cmd(cmd, &raft.ctx);
-
-                let result = response_rx.recv().await;
-                assert!(result.is_ok(), "Should receive response");
-
-                // **Assertion 2**: Linear read rejected with NOT_LEADER
-                if let Ok(Err(err)) = result {
-                    let err_str = format!("{:?}", err);
-                    assert!(
-                        err_str.contains("Not leader")
-                            || err_str.contains("NotLeader")
-                            || err_str.contains("NOT_LEADER")
-                            || err_str.contains("FailedPrecondition"),
-                        "Expected NOT_LEADER for linear read, got: {:?}",
-                        err
-                    );
-                } else {
-                    panic!("Linear read to Candidate should return NOT_LEADER error");
-                }
-            } else {
-                panic!("Expected Candidate state");
-            }
-        }
-
-        // **Test 3**: Eventual read should succeed (new behavior)
-        {
-            let (response_tx, mut response_rx) = MaybeCloneOneshot::new();
-            let read_req = ClientReadRequest {
-                client_id: 1,
-                keys: vec![bytes::Bytes::from("eventual_key")],
-                consistency_policy: Some(ReadConsistencyPolicy::EventualConsistency as i32),
-            };
-
-            if let RaftRole::Candidate(ref mut candidate) = raft.role {
-                let cmd = crate::ClientCmd::Read(read_req, response_tx);
-                candidate.push_client_cmd(cmd, &raft.ctx);
-
-                let result = response_rx.recv().await;
-                assert!(result.is_ok(), "Should receive response");
-
-                // **Assertion 3**: Eventual read succeeds on Candidate
-                if let Ok(Ok(read_response)) = result {
-                    // Check success_result for ReadResults
-                    assert!(
-                        read_response.success_result.is_some(),
-                        "Should have success_result"
-                    );
-                    if let Some(d_engine_proto::client::client_response::SuccessResult::ReadData(
-                        read_data,
-                    )) = read_response.success_result
-                    {
-                        assert!(!read_data.results.is_empty(), "Should have read results");
-                        assert_eq!(
-                            read_data.results[0].value,
-                            bytes::Bytes::from("eventual_value")
-                        );
-                    }
-                } else {
-                    panic!(
-                        "Eventual read to Candidate should succeed, got: {:?}",
-                        result
-                    );
-                }
-            } else {
-                panic!("Expected Candidate state");
-            }
-        }
-    }
-
-    // ========================================================================
-    // P1 PERFORMANCE VALIDATION TESTS - MEDIUM LOAD
-    // ========================================================================
-
-    // ========================================================================
-    // P1-1: MEDIUM LOAD - NATURAL BATCHING
-    // ========================================================================
-    //
-    // **Objective**: Verify adaptive batching under moderate load
-    //
-    // **Expected Behavior**:
-    // - Multiple drain cycles (not single batch)
-    // - Natural batching reflects real backlog (5-30 items per cycle typical)
-    // - System remains responsive
-    //
-    #[tokio::test]
-    async fn test_p1_medium_load_natural_batching() {
-        // **Setup**: Create Leader Raft instance
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup mocks for replication
-        let mut raft_log = crate::MockRaftLog::new();
-        raft_log.expect_last_entry_id().returning(|| 11);
-        raft_log.expect_flush().returning(|| Ok(()));
-        raft_log.expect_append_entries().returning(|_| Ok(()));
-        raft_log.expect_calculate_majority_matched_index().returning(|_, _, _| Some(11));
-        raft_log.expect_load_hard_state().returning(|| Ok(None));
-        raft_log.expect_save_hard_state().returning(|_| Ok(()));
-
-        let mut replication_handler = crate::MockReplicationCore::new();
-        replication_handler
-            .expect_handle_raft_request_in_batch()
-            .returning(|_, _, _, _, _| {
-                Ok(crate::AppendResults {
-                    commit_quorum_achieved: true,
-                    learner_progress: std::collections::HashMap::new(),
-                    peer_updates: std::collections::HashMap::new(),
-                })
-            });
-
-        raft.ctx.storage.raft_log = Arc::new(raft_log);
-        raft.ctx.handlers.replication_handler = replication_handler;
-
-        // Transition to Leader
-        raft.handle_role_event(crate::RoleEvent::BecomeCandidate)
-            .await
-            .expect("Should become Candidate");
-        raft.handle_role_event(crate::RoleEvent::BecomeLeader)
-            .await
-            .expect("Should become Leader");
-
-        // **Step 1**: Send ~50 commands with moderate spacing
-        const TOTAL_COMMANDS: usize = 50;
-        if let RaftRole::Leader(ref mut leader) = raft.role {
-            for i in 0..TOTAL_COMMANDS {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let write_cmd = WriteCommand {
-                    operation: Some(Operation::Insert(
-                        d_engine_proto::client::write_command::Insert {
-                            key: Bytes::from(format!("key_{}", i)),
-                            value: Bytes::from(format!("value_{}", i)),
-                            ttl_secs: 0,
-                        },
-                    )),
-                };
-                let write_req = ClientWriteRequest {
-                    client_id: 1,
-                    commands: vec![write_cmd],
-                };
-                let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-            }
-
-            // **Step 2**: Flush multiple times and verify batching behavior
-            let (role_tx, _role_rx) = tokio::sync::mpsc::unbounded_channel();
-            let mut flush_count = 0;
-
-            // Try flushing multiple times to drain all commands
-            for _ in 0..10 {
-                leader
-                    .flush_cmd_buffers(&raft.ctx, &role_tx)
-                    .await
-                    .expect("flush should succeed");
-                flush_count += 1;
-
-                // Small delay to simulate real load spacing
-                tokio::time::sleep(Duration::from_millis(1)).await;
-            }
-
-            // **Assertion**: Multiple flush cycles occurred
-            // (With moderate load, we expect more than 1 cycle but less than 50)
-            // This verifies natural batching behavior
-            assert!(flush_count > 0, "Should complete flush operations");
-        } else {
-            panic!("Expected Leader state");
-        }
-    }
-
-    // ========================================================================
-    // P1-2: MIXED WORKLOAD - READ/WRITE COEXISTENCE
-    // ========================================================================
-    //
-    // **Objective**: Verify different command types processed correctly
-    //
-    // **Expected Behavior**:
-    // - Single drain cycle collects all commands
-    // - Correct routing for different command types
-    // - All commands complete successfully
-    //
-    #[tokio::test]
-    async fn test_p1_mixed_workload_read_write_coexistence() {
-        // **Setup**: Create Leader Raft instance
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup mocks
-        let mut raft_log = crate::MockRaftLog::new();
-        raft_log.expect_last_entry_id().returning(|| 11);
-        raft_log.expect_flush().returning(|| Ok(()));
-        raft_log.expect_append_entries().returning(|_| Ok(()));
-        raft_log.expect_calculate_majority_matched_index().returning(|_, _, _| Some(11));
-        raft_log.expect_load_hard_state().returning(|| Ok(None));
-        raft_log.expect_save_hard_state().returning(|_| Ok(()));
-
-        let mut replication_handler = crate::MockReplicationCore::new();
-        replication_handler
-            .expect_handle_raft_request_in_batch()
-            .returning(|_, _, _, _, _| {
-                Ok(crate::AppendResults {
-                    commit_quorum_achieved: true,
-                    learner_progress: std::collections::HashMap::new(),
-                    peer_updates: std::collections::HashMap::new(),
-                })
-            });
-
-        raft.ctx.storage.raft_log = Arc::new(raft_log);
-        raft.ctx.handlers.replication_handler = replication_handler;
-
-        // Transition to Leader
-        raft.handle_role_event(crate::RoleEvent::BecomeCandidate)
-            .await
-            .expect("Should become Candidate");
-        raft.handle_role_event(crate::RoleEvent::BecomeLeader)
-            .await
-            .expect("Should become Leader");
-
-        // **Step 1**: Send mixed commands
-        if let RaftRole::Leader(ref mut leader) = raft.role {
-            // Send 10 write requests
-            for i in 0..10 {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let write_cmd = WriteCommand {
-                    operation: Some(Operation::Insert(
-                        d_engine_proto::client::write_command::Insert {
-                            key: Bytes::from(format!("write_key_{}", i)),
-                            value: Bytes::from(format!("write_value_{}", i)),
-                            ttl_secs: 0,
-                        },
-                    )),
-                };
-                let write_req = ClientWriteRequest {
-                    client_id: 1,
-                    commands: vec![write_cmd],
-                };
-                let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-            }
-
-            // Send 5 Eventual read requests
-            for i in 0..5 {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let read_req = ClientReadRequest {
-                    client_id: 1,
-                    keys: vec![Bytes::from(format!("read_key_{}", i))],
-                    consistency_policy: Some(ReadConsistencyPolicy::EventualConsistency as i32),
-                };
-                let cmd = crate::ClientCmd::Read(read_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-            }
-
-            // Send 5 Lease read requests (simpler than LinearizableRead)
-            for i in 0..5 {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let read_req = ClientReadRequest {
-                    client_id: 1,
-                    keys: vec![Bytes::from(format!("lease_key_{}", i))],
-                    consistency_policy: Some(ReadConsistencyPolicy::LeaseRead as i32),
-                };
-                let cmd = crate::ClientCmd::Read(read_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-            }
-
-            // **Step 2**: Flush all commands
-            let (role_tx, _role_rx) = tokio::sync::mpsc::unbounded_channel();
-            leader
-                .flush_cmd_buffers(&raft.ctx, &role_tx)
-                .await
-                .expect("flush should succeed");
-
-            // **Assertion**: All commands routed and processed successfully
-            // (Actual routing verification would require accessing internal state,
-            // but the absence of errors indicates successful processing)
-        } else {
-            panic!("Expected Leader state");
-        }
-    }
-
-    // ========================================================================
-    // P1-3: BURST LOAD - RECOVERY AFTER SPIKE
-    // ========================================================================
-    //
-    // **Objective**: Verify system recovers gracefully after traffic burst
-    //
-    // **Expected Behavior**:
-    // - Burst: Multiple drain cycles with max_batch_size
-    // - All burst requests processed
-    // - Trickle requests: each batch=1, no residual delay
-    // - System remains stable
-    //
-    #[tokio::test]
-    async fn test_p1_burst_load_recovery_after_spike() {
-        // **Setup**: Create Leader Raft instance
-        let (_graceful_tx, graceful_rx) = watch::channel(());
-        let mut raft = MockBuilder::new(graceful_rx).build_raft();
-
-        // Setup mocks
-        let mut raft_log = crate::MockRaftLog::new();
-        raft_log.expect_last_entry_id().returning(|| 11);
-        raft_log.expect_flush().returning(|| Ok(()));
-        raft_log.expect_append_entries().returning(|_| Ok(()));
-        raft_log.expect_calculate_majority_matched_index().returning(|_, _, _| Some(11));
-        raft_log.expect_load_hard_state().returning(|| Ok(None));
-        raft_log.expect_save_hard_state().returning(|_| Ok(()));
-
-        let mut replication_handler = crate::MockReplicationCore::new();
-        replication_handler
-            .expect_handle_raft_request_in_batch()
-            .returning(|_, _, _, _, _| {
-                Ok(crate::AppendResults {
-                    commit_quorum_achieved: true,
-                    learner_progress: std::collections::HashMap::new(),
-                    peer_updates: std::collections::HashMap::new(),
-                })
-            });
-
-        raft.ctx.storage.raft_log = Arc::new(raft_log);
-        raft.ctx.handlers.replication_handler = replication_handler;
-
-        // Transition to Leader
-        raft.handle_role_event(crate::RoleEvent::BecomeCandidate)
-            .await
-            .expect("Should become Candidate");
-        raft.handle_role_event(crate::RoleEvent::BecomeLeader)
-            .await
-            .expect("Should become Leader");
-
-        if let RaftRole::Leader(ref mut leader) = raft.role {
-            // **Step 1**: Burst load - 200 commands rapidly
-            const BURST_SIZE: usize = 200;
-            for i in 0..BURST_SIZE {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let write_cmd = WriteCommand {
-                    operation: Some(Operation::Insert(
-                        d_engine_proto::client::write_command::Insert {
-                            key: Bytes::from(format!("burst_key_{}", i)),
-                            value: Bytes::from(format!("burst_value_{}", i)),
-                            ttl_secs: 0,
-                        },
-                    )),
-                };
-                let write_req = ClientWriteRequest {
-                    client_id: 1,
-                    commands: vec![write_cmd],
-                };
-                let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-            }
-
-            // **Step 2**: Flush burst (multiple cycles expected)
-            let (role_tx, _role_rx) = tokio::sync::mpsc::unbounded_channel();
-            let mut burst_flushes = 0;
-            for _ in 0..20 {
-                leader
-                    .flush_cmd_buffers(&raft.ctx, &role_tx)
-                    .await
-                    .expect("flush should succeed");
-                burst_flushes += 1;
-                tokio::time::sleep(Duration::from_millis(1)).await;
-            }
-
-            // **Step 3**: Trickle load - 10 commands with spacing
-            const TRICKLE_SIZE: usize = 10;
-            for i in 0..TRICKLE_SIZE {
-                let (response_tx, _response_rx) = MaybeCloneOneshot::new();
-                let write_cmd = WriteCommand {
-                    operation: Some(Operation::Insert(
-                        d_engine_proto::client::write_command::Insert {
-                            key: Bytes::from(format!("trickle_key_{}", i)),
-                            value: Bytes::from(format!("trickle_value_{}", i)),
-                            ttl_secs: 0,
-                        },
-                    )),
-                };
-                let write_req = ClientWriteRequest {
-                    client_id: 1,
-                    commands: vec![write_cmd],
-                };
-                let cmd = crate::ClientCmd::Propose(write_req, response_tx);
-                leader.push_client_cmd(cmd, &raft.ctx);
-
-                // Flush after each trickle command
-                leader
-                    .flush_cmd_buffers(&raft.ctx, &role_tx)
-                    .await
-                    .expect("flush should succeed");
-
-                // Small delay between trickle requests
-                tokio::time::sleep(Duration::from_millis(10)).await;
-            }
-
-            // **Assertion**: System processed both burst and trickle successfully
-            assert!(burst_flushes > 0, "Burst should be processed");
-        } else {
-            panic!("Expected Leader state");
-        }
-    }
-}
